BZ 731863 - add --savedata (-s) option to aeolus-cleanup
https://bugzilla.redhat.com/show_bug.cgi?id=731863

When the -s option is used, the cleanup process will not remove any data

+++ recipes/aeolus/manifests/image-factory.pp
@@ -91,8 +91,10 @@ class aeolus::image-factory::disabled {
       hasstatus => true,
       enable  => false}
 
-  ### Destroy and cleanup aeolus artifacts
+  if $aeolus_save_data == "false" {
+    ### Destroy and cleanup aeolus artifacts
     exec{"remove_aeolus_templates":     command => "/bin/rm -rf /templates"}
+  }
 }
 
 define aeolus::image($template, $provider='', $target='', $hwp=''){
